requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
numpy
scikit-learn
requests
google-api-python-client
google-auth
google-auth-oauthlib
